Move 16x8 reference method to transpose_conv_common
