+++ tensorflow/lite/micro/kernels/transpose_conv.h
@@ -60,6 +60,9 @@ TfLiteStatus CalculateOpDataTransposeConv(
 
 TfLiteStatus TransposeConvPrepare(TfLiteContext* context, TfLiteNode* node);
 
+TfLiteStatus TransposeConvEvalInt16x8Reference(TfLiteContext* context,
+                                               TfLiteNode* node);
+
 // This is the most generic TfLiteRegistration. The actual supported types may
 // still be target dependent. The only requirement is that every implementation
 // (reference or optimized) must define this function.
